+++ lib/run.py
@@ -1,3 +1,3 @@
 import Uno
 
-Uno.main(int(input('How many players? ')))
+Uno.main()